Praveen-user-authorization(app): Updated the Users, Conferences, Sessions, Settings to support authentication (#20)
* Made User endpoints wokring with the user_id from the token

* Modified User Model to support Authorization

* Updated Users Model

* Updated Conference Model, Conference CRUD to support authorization

* Updated Sessions Model, Sessions CRUD to support authorization

* Updated Settings Models, Settings Crud to support authentication

+++ app/crud/sessions_crud.py
@@ -9,9 +9,8 @@ from pytz import timezone
 def get_sessions(db: Session, skip: int = 0, limit: int = 100):
     return db.query(models.Session).offset(skip).limit(limit).all()
 
-def create_conference_session(db: Session, session: schemas.SessionCreate, conference_id: int):
-    db_session = models.Session(**session.model_dump(), conference_id=conference_id)
-    owner_id=db.query(models.Conference).filter(models.Conference.id == conference_id).first().owner_id
+def create_conference_session(db: Session, session: schemas.SessionCreate, owner_id: int):
+    db_session = models.Session(**session.model_dump())
     tz = timezone('Asia/Kolkata')
     db_session.created_on = datetime.now(tz)
     db_session.updated_on = datetime.now(tz)
@@ -23,6 +22,10 @@ def create_conference_session(db: Session, session: schemas.SessionCreate, confe
 
 def get_session(db: Session, session_id: int):
     return db.query(models.Session).filter(models.Session.id == session_id).first()
+
+# get sessions by owner id and session id
+def get_session_by_owner_id(db: Session, session_id: int, owner_id: int):
+    return db.query(models.Session).filter(models.Session.id == session_id, models.Session.owner_id == owner_id).first()
 
 def get_sessions_by_name(db: Session, conference_id:int, name: str):
     return db.query(models.Session).filter(models.Session.conference_id == conference_id, models.Session.name == name).all()
@@ -52,12 +55,12 @@ def get_all_sessions_by_conference_id_between_date(db: Session, conference_id: i
     return db.query(models.Session).filter(models.Session.conference_id == conference_id, models.Session.date >= filter_start_date, models.Session.date <= filter_end_date).all()
 
 # get all conferences based on name or description strings
-def get_all_sessions_by_conference_id_and_name_or_description(db: Session, conference_id: int, search_string: str):
+def get_all_sessions_by_search(db: Session,conference_id:int, search_string: str):
     return db.query(models.Session).filter(models.Session.conference_id == conference_id, or_(models.Session.name.like('%'+search_string+'%'), models.Session.description.like('%'+search_string+'%'))).all()
 
 #delete session
-def delete_session(db: Session, session_id: int):
-    db.query(models.Session).filter(models.Session.id == session_id).delete()
+def delete_session(db: Session, session_id: int, owner_id: int):
+    db.query(models.Session).filter(models.Session.id == session_id,models.Session.owner_id == owner_id).delete()
     db.commit()
     return True
 
@@ -68,8 +71,8 @@ def delete_all_sessions_by_conference_id(db: Session, conference_id: int):
     return True
 
 #update session
-def update_session(db: Session, session: schemas.SessionCreate, session_id: int):
-    db_session = db.query(models.Session).filter(models.Session.id == session_id).first()
+def update_session(db: Session, session: schemas.SessionUpdate, session_id: int, owner_id: int):
+    db_session = db.query(models.Session).filter(models.Session.id == session_id,models.Session.owner_id == owner_id).first()
     tz = timezone('Asia/Kolkata')
     db_session.name = session.name
     db_session.date = session.date
